Split tree module by concern
